Add files via upload
My first assignment for this class wasn't too bad, it was a simple class used to calculate temperature averages. This was used using IntelliJ Community Version, coded with Java and utilized the Scanner + InputMismatchException (for simple user validation) to prevent crashes. Might've missed something, but too late to fix now.